perf: use clones-with-immutable-args for QTokens, solmate instead of OZ for the base ERC20 and ERC1155 implementations, remove clutter as in duplicate code and checks (#95)
* perf: optimize QToken storage layout, structs and function arguments

* build: update Certora QToken harnesses

* refactor: organize imports, interfaces and immutable variables

* fix(certora): CollateralToken interface in QuantCalculatorHarness

* feat: use clones-with-immutable-args for QTokens

also refactor and fix order of parameters on oracles and the PriceRegistry

* refactor: use Solmate for ERC1155 in the CollateralToken

also remove unused methods and events

* build: update forge-std dev dependency

* refactor: remove unnecessary, duplicate events

mints and burns can be distinguished and queried as transfers from and to the zero address

* refactor: update ERC1155 and CollateralToken usage and interfaces

* refactor: use a single method for getting predicted and existing tokens

* refactor: use staticcall with the identity precompile and remove unused methods

* feat: INVALID when input data > 128 bytes

also add documentation to the bytesToUint256Array function

* docs: natspec strings in OptionsUtils

* docs: add missing natspec string to the QToken

* refactor: final adjustments and missing documentation

* fix(ERC20): name and symbol string with zero bytes and domain separator

* refactor: annotate assembly blocks as memory-safe

* fix: QToken import

* test: fix Hardhat tests

* build: update Hardhat cache that was conflicting with Foundry's

* fix: Certora FundsCalculator library wrapper

* build: update Certora harnesses

* build: update dependencies

* fix: Certora collateralToken and fundsCalculator specs

* feat: add assetsRegistry and priceRegistry to IQuantCalculator

* fix: Certora specs and harnesses

* build: update clones-with-immutable-args dep

* style(lint): ignore contracts with 'memory-safe' annotations

* build: fix Certora validQToken mintSpread rule

* build(coverage): ignore files with the memory-safe annotation

* build: skip coverage due to unoptimized compilation issues

* build: update Solidity version to 0.8.14

* build: update solmate and format solcover.js

* build: update clones-with-immutable-args dep

* perf: store string length as the last byte in the uint array

* refactor: use solmate for safe transfers and ReentrancyGuard

* perf: save one mload and xor when storing the string length

* Revert "perf: save one mload and xor when storing the string length"

This reverts commit 863304d602b5b4b3221a80d63aa896dcf290e873.

* build: fix Certora ERC20 imports

* build: fix Certora Controller spec

* build: update clones-with-immutable-args dep

* test: fix custom error expectation and remove unnecessary library linking

* build: update Certora FundsCalculator spec run script

* feat: add custom errors for inline assembly code

also further optimize the OptionsUtils library and the QuantCalculator contract

* feat: expose non-reentrant single Controller actions

* build: fix Yarn lockfile

* build: update Certora FundsCalculator spec run script

* Revert "build: update Certora FundsCalculator spec run script"

This reverts commit db29ea90de76438077d75dd28dc96e9dfc2064f7.

* build: update Certora FundsCalculator spec run script

* refactor(erc20): read name and symbol strings directly from calldata

* refactor: simplify creation of immutable arg strings with 128 bytes

* build: remove commented code from Certora specs

* refactor: explicit modifiers for the option and spread creation functions

* test: rewrite broken test case using Forge

* test: rewrite broken overflow test case using Forge

* build: remove unused CustomErrors file

* build: update Solidity version to 0.8.15

* build(certora): fix broken and comment inapplicable CollateralToken rules

+++ contracts/interfaces/IChainlinkOracleManager.sol
@@ -10,7 +10,7 @@ interface IChainlinkOracleManager is
 {
     event PriceRegistrySubmission(
         address asset,
-        uint256 expiryTimestamp,
+        uint88 expiryTimestamp,
         uint256 price,
         uint256 expiryRoundId,
         address priceSubmitter,
@@ -23,7 +23,7 @@ interface IChainlinkOracleManager is
     /// @param _roundIdAfterExpiry the chainlink round id immediately after the expiry timestamp
     function setExpiryPriceInRegistryByRound(
         address _asset,
-        uint256 _expiryTimestamp,
+        uint88 _expiryTimestamp,
         uint256 _roundIdAfterExpiry
     ) external;
 
@@ -33,7 +33,7 @@ interface IChainlinkOracleManager is
     /// @param _asset address of asset to search price for
     /// @param _expiryTimestamp expiry timestamp to find the price at or before
     /// @return the round id immediately after the timestamp submitted
-    function searchRoundToSubmit(address _asset, uint256 _expiryTimestamp)
+    function searchRoundToSubmit(address _asset, uint88 _expiryTimestamp)
         external
         view
         returns (uint80);
